cannot add same favorite...cannot remove non-existent favorite implemented

+++ src/app/components/author-item/author-item.component.ts
@@ -19,18 +19,28 @@ export class AuthorItemComponent implements OnInit {
     this.author.favorite = true
     let favorites = JSON.parse(localStorage.getItem('favorites'))
     if(favorites == null) favorites = [];
-    favorites.unshift(author)
-    localStorage.setItem("favorites", JSON.stringify(favorites));
+    let exists = favorites.find(({ _id }) => author._id === _id)
+    if(!exists) {
+      favorites.unshift(author)
+      localStorage.setItem("favorites", JSON.stringify(favorites));
+    } else {
+      alert('This author is already in your favorites')
+    }
   }
 
   // Remove from favorite
   removeFavorite(author) {
     this.author.favorite = false
     let favorites = JSON.parse(localStorage.getItem('favorites'))
-    favorites = favorites.filter((item) => {
-      return item._id !== author._id
-    })
-    localStorage.setItem("favorites", JSON.stringify(favorites));
+    let exists = favorites.find(({ _id }) => author._id === _id)
+    if(exists) {
+      favorites = favorites.filter((item) => {
+        return item._id !== author._id
+      })
+      localStorage.setItem("favorites", JSON.stringify(favorites));
+    } else {
+      alert('This author does not belong to the favorites list')
+    }
   }
 
 }
